feat(marketplace): merge Roles into the Marketplace as one hub surface
The store (Browse) and the familiars' setup (Roles / Skills /
Capabilities) now live on a single Marketplace page with an accessible
section tablist, per-section hero copy/stats/search, and a "Your setup"
cross-nav group in the store's category rail. The standalone Roles page
(PluginsView) is retired; its role cards move to
marketplace/roles-section.tsx. The roles/capabilities workspace modes
survive as deep links that open the hub on the matching section, and the
roles add-on gate is removed from the sidebar, palette, and settings.
