+++ src/components/command-palette.tsx
@@ -569,7 +569,6 @@ export function CommandPalette({
           if (fm.id === "terminal") return addons?.terminal === true;
           if (fm.id === "browser") return addons?.browser === true;
           if (fm.id === "flow") return addons?.flow === true;
-          if (fm.id === "roles") return addons?.roles === true;
           if (fm.id === "groupchat") return addons?.groupchat === true;
           if (fm.id === "journal") return addons?.journal === true;
           return true;
